refactor: move github integration JS into collavre_github engine
Previously github_integration.js lived in the host app's app/javascript/,
so other host apps had to copy the file to enable the markdown sync UI.
Move it into the engine alongside collavre_slack/notion/plan, register
it in the esbuild GEM_ENGINES list, and import it from the engine path
in the host application.js.
