+++ app/javascript/application.js
@@ -18,10 +18,10 @@ import "collavre"
 import "collavre_slack"
 import "collavre_notion"
 import "collavre_plan"
+import "collavre_github"
 
 // Host app specific modules
 import "./firebase_config"
-import "./github_integration"
 import "./timezone_detection"
 import "./oauth_callback"
 import "./doorkeeper_token"
